* Check that a player only creates one private channel * Change unsigned long to uint32_t in autoid * Remove compiler warnings

+++ otserv/trunk/creature.cpp
@@ -37,7 +37,7 @@
 #include <vector>
 
 OTSYS_THREAD_LOCKVAR AutoID::autoIDLock;
-unsigned long AutoID::count = 1000;
+uint32_t AutoID::count = 1000;
 AutoID::list_type AutoID::list;
 
 extern Game g_game;
