+++ otserv/trunk/templates.h
@@ -41,11 +41,11 @@ public:
 		list[t->getID()] = t;
 	}
 	
-	void removeList(unsigned long _id){
+	void removeList(uint32_t _id){
 		list.erase(_id);
 	}
 	
-	typedef std::map<unsigned long, T*> list_type;
+	typedef std::map<uint32_t, T*> list_type;
 	list_type list;
 
 	typedef typename list_type::iterator listiterator;
@@ -74,13 +74,13 @@ public:
 			list.erase(it);
 	}
 
-	typedef std::set<unsigned long> list_type;
+	typedef std::set<uint32_t> list_type;
 
-	unsigned long auto_id;
+	uint32_t auto_id;
 	static OTSYS_THREAD_LOCKVAR autoIDLock;
 	
 protected:
-	static unsigned long count;
+	static uint32_t count;
 	static list_type list;
 
 };
